refactor: refatora estrutura de pastas, aplicando DDD para pastas do front
