+++ src/app/clients/actions/clients-actions.ts
@@ -1,41 +1,41 @@
 "use server";
 
 import {
-  serviceDataGET,
-  serviceDataAppend,
-  serviceDataById,
-  serviceDataUpdate,
-  serviceDataDelete,
-} from "@/modules/clients/services";
-import { mapSheetDataToClients } from "@/modules/clients/utils";
-import { ClientInput } from "@/modules/clients/schema";
+  dataGETService,
+  dataAppendService,
+  dataByIdService,
+  dataUpdateService,
+  dataDeleteService,
+} from "@/modules/clients/clients-services";
+import { mapSheetDataToClients } from "@/modules/clients/clients-utils";
+import { ClientInput } from "@/modules/clients/clients-schema";
 import { revalidatePath } from "next/cache";
 
 export async function getClientsAction() {
-  const sheetData = await serviceDataGET();
+  const sheetData = await dataGETService();
   const clients = mapSheetDataToClients(sheetData);
   return clients;
 }
 
 export async function getClientByIdAction(clientId: string) {
-  const sheetData = await serviceDataById(clientId);
+  const sheetData = await dataByIdService(clientId);
   return sheetData;
 }
 
 export async function appendClientsAction(data: ClientInput) {
-  const sheetData = await serviceDataAppend(data);
+  const sheetData = await dataAppendService(data);
 
   revalidatePath("/clients");
 }
 
 export async function updateClientsAction(data: ClientInput) {
-  const sheetData = await serviceDataUpdate(data);
+  const sheetData = await dataUpdateService(data);
   revalidatePath("/clients");
   return sheetData;
 }
 
 export async function deleteClientsAction(data: ClientInput) {
-  const sheetData = await serviceDataDelete(data);
+  const sheetData = await dataDeleteService(data);
   revalidatePath("/clients");
   return sheetData;
 }
