+++ src/app/clients/components/UI/clients-table.tsx
