+++ src/app/clients/components/UI/modal-client.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { ClientInput } from "@/modules/clients/schema";
+import { ClientInput } from "@/modules/clients/clients-schema";
 import { ClientForm } from "./form-client";
 import { X } from "lucide-react";
