+++ src/app/signin/components/UI/signin-form.tsx
@@ -1,4 +1,4 @@
-import { login } from "@/app/(public)/signin/actions/signin-action";
+import { login } from "@/app/signin/actions/signin-action";
 
 export default function SignInForm() {
   return (
