+++ src/app/signout/actions/logout-action.ts
